fasm: add bel pins features handling
- added CARRY4 FASM output
- moved lut mapper to another file
